+++ src/main/kotlin/krangl/Select.kt
@@ -1,69 +1,46 @@
 package krangl
 
-
-//
-// DataFrame extensions
-//
+/** Extension API to allow for column subsetting (both positive and negative)*/
 
 /**
  * @author Holger Brandl
  */
 
-/** Select columns by predicate.
-
-Example
-
-```
-foo.select{ it is IntCol }
-foo.select{ it.name.startsWith("bar") }
-```
- */
-//fun DataFrame.select(colSelector: (DataCol) -> Boolean) = selectByName(cols.filter(colSelector).map{it.name})
-//
-//
-///** Convenience wrapper around to work with varag <code>krangl.DataFrame.select</code> */
-//fun DataFrame.selectByName(columns: List<String>): DataFrame = selectByName(*columns.toTypedArray())
-//
-///** Keeps only the variables that match any of the given expressions. E.g. use `startsWith("foo")` to select for columns staring with 'foo'.*/
-//fun DataFrame.selectByName(which: ColNames.() -> List<Boolean?>): DataFrame = selectByName(*arrayOf(which))
-//
-//fun DataFrame.selectByName(vararg which: ColNames.() -> List<Boolean?>): DataFrame {
-//    val reducedSelector = reduceColSelectors(which)
-//
-//    return select(reducedSelector)
-//}
-
 
 //
-// select() support API
+// Public API
 //
 
+
+
+//https://kotlinlang.org/docs/reference/inline-functions.html#reified-type-parameters
+/** Select columns by column type */
+inline fun <reified T : DataCol> DataFrame.select() = select(cols.filter { it is T }.map { it.name })
+
+/** Remove columns by column type */
+inline fun <reified T : DataCol> DataFrame.remove() = select(cols.filter { !(it is T) }.map { it.name })
+//inline fun <reified T : DataCol> DataFrame.select() = select(cols.filter{ it is T }.map{it.name})
+
+
+class InvalidColumnSelectException(msg: String) : RuntimeException(msg)
 
 class ColNames(val names: List<String>)
 
-// select utitlies see http://www.rdocumentation.org/packages/dplyr/functions/select
+typealias ColumnSelector = ColNames.() -> List<Boolean?>
+
+//
+// select utilities inspired by see http://www.rdocumentation.org/packages/dplyr/functions/select
+//
+
 fun ColNames.matches(regex: String) = names.map { it.matches(regex.toRegex()) }
 
-// todo add possiblity to negative selection with mini-language. E.g. -startsWith("name")
+fun ColNames.startsWith(prefix: String) = names.map { it.startsWith(prefix) }.falseAsNull()
+fun ColNames.endsWith(prefix: String) = names.map { it.endsWith(prefix) }.falseAsNull()
+fun ColNames.matches(regex: Regex) = names.map { it.matches(regex) }.falseAsNull()
 
-internal fun List<Boolean>.falseAsNull() = map { if (!it) null else true }
-internal fun List<Boolean>.trueAsNull() = map { if (it) null else false }
-internal fun List<Boolean?>.nullAsFalse(): List<Boolean> = map { it ?: false }
-
-fun ColNames.startsWith(prefix: String) = names.map { it.startsWith(prefix) }
-fun ColNames.endsWith(prefix: String) = names.map { it.endsWith(prefix) } //.falseAsNull()
-//fun ColNames.everything() = Array(names.size, { true }).toList() // unclear purpose
-fun ColNames.matches(regex: Regex) = names.map { it.matches(regex) } //.falseAsNull()
-
-fun ColNames.oneOf(vararg someNames: String): List<Boolean?> = names.map { someNames.contains(it) } //.falseAsNull()
-fun ColNames.oneOf(someNames: List<String>): List<Boolean?> = names.map { someNames.contains(it) } //.falseAsNull()
-
-// normallthere should be no need for them. We just do positive selection and either use renmove or select
-// BUT: verbs like gather still need to support negative selection
-internal fun ColNames.except(vararg columns: String) = names.map { !columns.contains(it) } //.falseAsNull()
-
-internal fun ColNames.except(columnSelector: ColumnSelector) = columnSelector(this).not()
-
+fun ColNames.oneOf(vararg someNames: String): List<Boolean?> = names.map { someNames.contains(it) }.falseAsNull()
+fun ColNames.oneOf(someNames: List<String>): List<Boolean?> = names.map { someNames.contains(it) }.falseAsNull()
+fun ColNames.all() = Array(names.size, { true }).toList() // unclear purpose
 
 fun ColNames.range(from: String, to: String): List<Boolean?> {
     val rangeStart = names.indexOf(from)
@@ -74,31 +51,71 @@ fun ColNames.range(from: String, to: String): List<Boolean?> {
 }
 
 
-// since this affects String namespace it might be not a good idea
-@Deprecated("will be removed since this affects String namespace it might be not a good idea")
-operator fun String.unaryMinus() = fun ColNames.(): List<Boolean?> = names.map { it != this@unaryMinus }.trueAsNull()
+fun ColumnSelector.unaryNot(): ColumnSelector = fun ColNames.(): List<Boolean?> = this.this@unaryNot().map{ it?.not() }
+
+
+// normally, there should be no need for them. We just do positive selection and either use renmove or select
+// BUT: verbs like gather still need to support negative selection
+fun ColNames.except(vararg columns: String) = names.map { !columns.contains(it) }.trueAsNull()
+fun ColNames.except(columnSelector: ColumnSelector) = columnSelector(this).not()
+
+
+
+
+// commented out because it's not clear how to use it
+//val foo: ColumnSelector = { startsWith("foo")
+//sleepData.select(foo AND { endsWith("dfd") })
+//infix fun ColumnSelector.AND(other: ColumnSelector): ColumnSelector = fun ColNames.(): List<Boolean?> {
+//    return this.this@AND().zip(this.other()).map { nullAwareAnd(it.first, it.second) }
+//}
+
+
+//@Deprecated("will be removed since this affects String namespace it might be not a good idea")
+//operator fun String.unaryMinus() = fun ColNames.(): List<Boolean?> = names.map { it != this@unaryMinus }.trueAsNull()
 //operator fun Iterable<String>.unaryMinus() = fun ColNames.(): List<Boolean> = names.map { !this@unaryMinus.contains(it) }
 //operator fun List<Boolean?>.unaryMinus() = not()
 
-fun List<Boolean?>.not() = map { it?.not() }
-//operator fun List<Boolean?>.not() = map { it?.not() } // todo needed?
+operator fun List<Boolean?>.not() = map { it?.not() }
 
-//val another = -"dsf"
+
+//
+// Internal API
+//
+
 
 internal fun DataFrame.reduceColSelectors(which: Array<out ColumnSelector>): List<Boolean?> =
         which.map { it(ColNames(names)) }.reduce { selA, selB -> selA nullAwareAND selB }
 
 
-private infix fun List<Boolean?>.nullAwareAND(other: List<Boolean?>): List<Boolean?> = this.zip(other).map {
-    it.run {
-        if (first == null && second == null) {
-            null
-        } else if (first != null && second != null) {
-            20
-            first!! && second!!
-        } else {
-            first ?: second
-        }
+internal infix fun List<Boolean?>.nullAwareAND(other: List<Boolean?>): List<Boolean?> = this.zip(other).map {
+    nullAwareAnd(it.first, it.second)
+}
+
+
+internal fun List<Boolean>.falseAsNull() = map { if (!it) null else true }
+internal fun List<Boolean>.trueAsNull() = map { if (it) null else false }
+internal fun List<Boolean?>.nullAsFalse(): List<Boolean> = map { it ?: false }
+
+
+
+// todo the collapse logic does not seem right: why would would null && true be true?
+internal fun nullAwareAnd(first: Boolean?, second: Boolean?): Boolean? {
+    return if (first == null && second == null) {
+        null
+    } else if (first != null && second != null) {
+        first && second
+    } else {
+        first ?: second
+    }
+}
+
+internal fun nullAwareOr(first: Boolean?, second: Boolean?): Boolean? {
+    return if (first == null && second == null) {
+        null
+    } else if (first != null && second != null) {
+        first || second
+    } else {
+        first ?: second
     }
 }
 
@@ -122,7 +139,3 @@ internal fun DataFrame.colSelectAsNames(which: List<Boolean?>): List<String> {
     return colSelection
 }
 
-
-//https://kotlinlang.org/docs/reference/inline-functions.html#reified-type-parameters
-inline fun <reified T : DataCol> DataFrame.select() = select(cols.filter { it is T }.map { it.name })
-//inline fun <reified T : DataCol> DataFrame.select() = select(cols.filter{ it is T }.map{it.name})